Move home page plugins into home-page folder.
