+++ website/static/js/home-page/quickProjectSearchPlugin.js
