+++ website/static/js/home-page/recentActivityWidget.js
